requirements.txt
-i https://pypi.org/simple
django-cors-headers==3.2.0
django==2.2.8
djangorestframework-jwt==1.11.0
djangorestframework==3.10.3
pyjwt==1.7.1
pytz==2019.3
sqlparse==0.3.0
